combine create.go update.go into input.go
